address PR comments x1
