+++ src/Aspire.Hosting/Exec/IContainerExecService.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Aspire.Hosting.Exec;
+
+/// <summary>
+/// A service to execute container exec commands.
+/// </summary>
+public interface IContainerExecService
+{
+
+}
