+++ tests/Aspire.Hosting.Tests/Utils/TestDcpExecutor.cs
@@ -18,5 +18,5 @@ internal sealed class TestDcpExecutor : IDcpExecutor
     public Task StopResourceAsync(IResourceReference resourceReference, CancellationToken cancellationToken) => Task.CompletedTask;
 
     public Task<AppResource> RunEphemeralResourceAsync(IResource ephemeralResource, CancellationToken cancellationToken) => throw new NotImplementedException();
-    public void DeleteEphemeralResource(AppResource? ephemeralResource) { }
+    public Task DeleteEphemeralResourceAsync(AppResource ephemeralResource) => Task.CompletedTask;
 }
